We are gonna add some sections to the Leading Technical Change page. Section 1 is gonna have title "What is LTC? - “Better.” and body is going to be "How can we help our software teams get stronger? Faster? Smarter? Happier? More positive and more productive?
There are lots of ideas about what changes we might desire to make, ranging from process to policy to procedure to skillset. There’s no problem in finding proposed changes. No, the problem is in helping your team to actually make any given change.

Leading Technical Change (LTC), is a course designed to focus on how to make change, not which change to make.

Download the LTC Overview for your team lead or department head today! 👇" Underneath this have the download course outline button. This will be on the left and have a big 16:9 placeholder image on the right.