Let's add another section. This section will have the title "Theory of Change" and the body will be "Many people, conceiving of some change they want to make, focus all of their attention only on the “what” of that particular change.  But successful change, *any* successful change, depends at least as much on the “how” as on the “what”.
The  theory of change we’ll be working with in the course is based around research on “cognitive frames”, the short-hand pre-assembled bits and pieces that actually underlie most day-to-day action. We can’t eliminate frames (and we wouldn’t want to), we can only change their shape.

Doing that involves three abstract strategies:

Take many more much smaller steps (MMMSS).
Optimize for the humans (OFTH).
Make Change Normal (MCN).
From these three broad strategies, we can develop specific, powerful tactics to increase the likelihood that our changes will really work, and really stick." and this one is on the left with two 16:9 images on the right.